Refactor logic. Divide in smaller functions
